backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database and ORM
sqlmodel==0.0.14
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# Authentication and JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration and environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code quality
ruff==0.1.14
black==24.1.1
